refactor(gamma-platform): move the organization model and service to shared
The endpoint serves the whole organization entity, not a platform-policies projection; shared/services already holds its console sub-resource.
