+++ gravitee-gamma/gravitee-gamma-module-platform/src/main/ui/features/platform-policies/utils/platformFlowAdapter.ts
@@ -15,8 +15,8 @@
  */
 import type { ConditionSelector, Flow, FlowExecution, HttpMethod, HttpSelector, Selector, Step } from '@gravitee/graphene-policy-studio';
 
+import type { PlatformFlow, PlatformFlowMode, PlatformFlowStep } from '../../../shared/types/organization';
 import { toStudioStep } from '../../../shared/v2-flows';
-import type { PlatformFlow, PlatformFlowMode, PlatformFlowStep } from '../types/platformPolicies';
 
 /** The organization flow schema always carries a path operator; a flow saved without one would never match. */
 const DEFAULT_PATH_OPERATOR = { path: '/', operator: 'STARTS_WITH' } as const;
