+++ gravitee-gamma/gravitee-gamma-module-platform/src/main/ui/shared/services/organization.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { apimFetchJsonOrg } from '../../../shared/api/apimClient';
-import type { Organization } from '../types/platformPolicies';
+import { apimFetchJsonOrg } from '../api/apimClient';
+import type { Organization } from '../types/organization';
 
 /** Platform flows are stored on the organization entity itself — there is no dedicated flows endpoint. */
 export function getOrganization(): Promise<Organization> {
@@ -26,7 +26,10 @@ export function getOrganization(): Promise<Organization> {
  * Full-entity update: every field read from the organization has to be sent back. The resource binds the
  * body to `UpdateOrganizationEntity`, which is `OrganizationEntity` without `id`; the extra `id` we send
  * is ignored, the mapper has `FAIL_ON_UNKNOWN_PROPERTIES` disabled.
+ *
+ * The resource answers `204`, so there is no saved entity to read back: callers that need the stored
+ * state have to re-read the organization.
  */
-export function updateOrganization(organization: Organization): Promise<Organization> {
-    return apimFetchJsonOrg<Organization>('', { method: 'PUT', body: JSON.stringify(organization) });
+export function updateOrganization(organization: Organization): Promise<void> {
+    return apimFetchJsonOrg<void>('', { method: 'PUT', body: JSON.stringify(organization) });
 }
